refactor(BA-6191): move ScopeContext to base types module
ScopeContext is a shared scope locator used by both scope creation and
teardown; defining it in scope_creator forced scope_purger to import from
scope_creator. Move it to repositories/base/types.py so both import it from a
neutral, model-light module.
